fix(both datsetHash() and hashOfHash() returns integer)
py4j maps int to Java.lang.Integer, which can break API that accepts
Java.lang.Long only.  So we will always use an Integer when passing
numeric values from Python to JVM.

+++ src/main/java/org/tresamigos/smv/SmvDataSetRepository.java
@@ -96,5 +96,5 @@ public interface SmvDataSetRepository {
 	 * the hash for all its super classes up to and excluding the base
 	 * class provided by SMV.
 	 */
-	long datasetHash(String modfqn, boolean includeSuperClass);
+	int datasetHash(String modfqn, boolean includeSuperClass);
 }
